+++ src/RSSFeed.php
@@ -38,26 +38,6 @@ use Glpi\RichText\RichText;
 use Glpi\Toolbox\URL;
 use SimplePie\SimplePie;
 
-// $feed = new SimplePie();
-// $feed->set_cache_duration(3600);
-// $feed->set_feed_url('http://linuxfr.org/news.atom');
-// $feed->force_feed(true);
-// // Initialize the whole SimplePie object.  Read the feed, process it, parse it, cache it, and
-// // all that other good stuff.  The feed's information will not be available to SimplePie before
-// // this is called.
-// $success = $feed->init();
-//
-// // We'll make sure that the right content type and character encoding gets set automatically.
-// // This function will grab the proper character encoding, as well as set the content type to text/html.
-// $feed->handle_content_type();
-// if ($feed->error())
-// {
-//    echo "ERROR";
-// } else {
-//    echo $feed->get_title();
-//    echo $feed->get_link();
-// }
-
 /**
  * RSSFeed Class
  *
@@ -70,57 +50,46 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
     public static $rightname    = 'rssfeed_public';
 
-    const PERSONAL = 128;
+    public const PERSONAL = 128;
 
     public static function getTypeName($nb = 0)
     {
-
         if (Session::haveRight('rssfeed_public', READ)) {
             return _n('RSS feed', 'RSS feed', $nb);
         }
         return _n('Personal RSS feed', 'Personal RSS feed', $nb);
     }
 
-
     public static function canCreate()
     {
-
         return (Session::haveRightsOr(self::$rightname, [CREATE, self::PERSONAL]));
     }
 
-
     public static function canView()
     {
-
         return (Session::haveRightsOr(self::$rightname, [READ, self::PERSONAL]));
     }
 
-
     public function canViewItem()
     {
-
-       // Is my rssfeed or is in visibility
-        return (($this->fields['users_id'] == Session::getLoginUserID())
+        // Is my rssfeed or is in visibility
+        return (($this->fields['users_id'] === Session::getLoginUserID())
               || (Session::haveRight('rssfeed_public', READ)
                   && $this->haveVisibilityAccess()));
     }
 
-
     public function canCreateItem()
     {
-       // Is my rssfeed
-        return ($this->fields['users_id'] == Session::getLoginUserID());
+        // Is my rssfeed
+        return ($this->fields['users_id'] === Session::getLoginUserID());
     }
-
 
     public function canUpdateItem()
     {
-
-        return (($this->fields['users_id'] == Session::getLoginUserID())
+        return (($this->fields['users_id'] === Session::getLoginUserID())
               || (Session::haveRight('rssfeed_public', UPDATE)
                   && $this->haveVisibilityAccess()));
     }
-
 
     /**
      * @since 0.85
@@ -131,7 +100,6 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return (Session::haveRightsOr(self::$rightname, [UPDATE, self::PERSONAL]));
     }
 
-
     /**
      * @since 0.85
      * for personal rss feed
@@ -141,44 +109,30 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return (Session::haveRightsOr(self::$rightname, [PURGE, self::PERSONAL]));
     }
 
-
-    /**
-     * @since 0.85
-     *
-     * @see CommonDBTM::canPurgeItem()
-     **/
     public function canPurgeItem()
     {
-
-        return (($this->fields['users_id'] == Session::getLoginUserID())
+        return (($this->fields['users_id'] === Session::getLoginUserID())
               || (Session::haveRight(self::$rightname, PURGE)
                   && $this->haveVisibilityAccess()));
     }
 
-
     public function post_getFromDB()
     {
-
-       // Users
+        // Users
         $this->users    = RSSFeed_User::getUsers($this->fields['id']);
 
-       // Entities
+        // Entities
         $this->entities = Entity_RSSFeed::getEntities($this->fields['id']);
 
-       // Group / entities
+        // Group / entities
         $this->groups   = Group_RSSFeed::getGroups($this->fields['id']);
 
-       // Profile / entities
+        // Profile / entities
         $this->profiles = Profile_RSSFeed::getProfiles($this->fields['id']);
     }
 
-
-    /**
-     * @see CommonDBTM::cleanDBonPurge()
-     **/
     public function cleanDBonPurge()
     {
-
         $this->deleteChildrenAndRelationsFromDb(
             [
                 Entity_RSSFeed::class,
@@ -196,58 +150,6 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         }
 
         return parent::haveVisibilityAccess();
-    }
-
-    /**
-     * Return visibility joins to add to SQL
-     *
-     * @param $forceall force all joins (false by default)
-     *
-     * @return string joins to add
-     **/
-    public static function addVisibilityJoins($forceall = false)
-    {
-       //not deprecated because used in Search
-        /** @var \DBmysql $DB */
-        global $DB;
-
-       //get and clean criteria
-        $criteria = self::getVisibilityCriteria();
-        unset($criteria['WHERE']);
-        $criteria['FROM'] = self::getTable();
-
-        $it = new \DBmysqlIterator(null);
-        $it->buildQuery($criteria);
-        $sql = $it->getSql();
-        $sql = trim(str_replace(
-            'SELECT * FROM ' . $DB->quoteName(self::getTable()),
-            '',
-            $sql
-        ));
-        return $sql;
-    }
-
-
-    /**
-     * Return visibility SQL restriction to add
-     *
-     * @return string restrict to add
-     **/
-    public static function addVisibilityRestrict()
-    {
-       //not deprecated because used in Search
-
-       //get and clean criteria
-        $criteria = self::getVisibilityCriteria();
-        unset($criteria['LEFT JOIN']);
-        $criteria['FROM'] = self::getTable();
-
-        $it = new \DBmysqlIterator(null);
-        $it->buildQuery($criteria);
-        $sql = $it->getSql();
-        $sql = preg_replace('/.*WHERE /', '', $sql);
-
-        return $sql;
     }
 
     /**
@@ -271,8 +173,8 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
             ];
         }
 
-       //JOINs
-       // Users
+        // JOINs
+        // Users
         $join['glpi_rssfeeds_users'] = [
             'ON' => [
                 'glpi_rssfeeds_users'   => 'rssfeeds_id',
@@ -288,7 +190,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         ];
         $orwhere = [];
 
-       // Groups
+        // Groups
         if (
             $forceall
             || (isset($_SESSION["glpigroups"]) && count($_SESSION["glpigroups"]))
@@ -314,11 +216,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         }
 
        // Profiles
-        if (
-            $forceall
-            || (isset($_SESSION["glpiactiveprofile"])
-              && isset($_SESSION["glpiactiveprofile"]['id']))
-        ) {
+        if ($forceall || isset($_SESSION["glpiactiveprofile"]['id'])) {
             $join['glpi_profiles_rssfeeds'] = [
                 'ON' => [
                     'glpi_profiles_rssfeeds'   => 'rssfeeds_id',
@@ -327,7 +225,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
             ];
         }
 
-        if (isset($_SESSION["glpiactiveprofile"]) && isset($_SESSION["glpiactiveprofile"]['id'])) {
+        if (isset($_SESSION["glpiactiveprofile"]['id'])) {
             $restrict = getEntitiesRestrictCriteria('glpi_entities_rssfeeds', '', '', true);
             if (!count($restrict)) {
                 $restrict = [true];
@@ -343,7 +241,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
             ];
         }
 
-       // Entities
+        // Entities
         if (
             $forceall
             || (isset($_SESSION["glpiactiveentities"]) && count($_SESSION["glpiactiveentities"]))
@@ -373,14 +271,8 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return $criteria;
     }
 
-    /**
-     * @param $field
-     * @param $values
-     * @param $options   array
-     **/
     public static function getSpecificValueToDisplay($field, $values, array $options = [])
     {
-
         if (!is_array($values)) {
             $values = [$field => $values];
         }
@@ -391,16 +283,8 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return parent::getSpecificValueToDisplay($field, $values, $options);
     }
 
-
-    /**
-     * @param $field
-     * @param $name               (default '')
-     * @param $values             (default '')
-     * @param $options      array
-     **/
     public static function getSpecificValueToSelect($field, $name = '', $values = '', array $options = [])
     {
-
         if (!is_array($values)) {
             $values = [$field => $values];
         }
@@ -413,7 +297,6 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return parent::getSpecificValueToSelect($field, $name, $values, $options);
     }
 
-
     public function rawSearchOptions()
     {
         $tab = [];
@@ -425,7 +308,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '1',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'name',
             'name'               => __('Name'),
             'datatype'           => 'itemlink',
@@ -445,7 +328,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '3',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'url',
             'name'               => __('URL'),
             'datatype'           => 'string',
@@ -454,7 +337,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '4',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'is_active',
             'name'               => __('Active'),
             'datatype'           => 'bool',
@@ -463,7 +346,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '6',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'have_error',
             'name'               => _n('Error', 'Errors', 1),
             'datatype'           => 'bool',
@@ -472,7 +355,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '7',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'max_items',
             'name'               => __('Number of items displayed'),
             'datatype'           => 'number',
@@ -485,7 +368,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '16',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'comment',
             'name'               => __('Comments'),
             'datatype'           => 'text'
@@ -493,7 +376,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '5',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'refresh_rate',
             'name'               => __('Refresh rate'),
             'datatype'           => 'timestamp',
@@ -513,7 +396,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '19',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'date_mod',
             'name'               => __('Last update'),
             'datatype'           => 'datetime',
@@ -522,7 +405,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         $tab[] = [
             'id'                 => '121',
-            'table'              => $this->getTable(),
+            'table'              => static::getTable(),
             'field'              => 'date_creation',
             'name'               => __('Creation date'),
             'datatype'           => 'datetime',
@@ -535,17 +418,12 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return $tab;
     }
 
-
-    /**
-     * @see CommonGLPI::getTabNameForItem()
-     **/
     public function getTabNameForItem(CommonGLPI $item, $withtemplate = 0)
     {
-
         if (self::canView()) {
             $nb = 0;
-            switch ($item->getType()) {
-                case 'RSSFeed':
+            switch ($item::class) {
+                case RSSFeed::class:
                     $showtab = [1 => self::createTabEntry(__('Content'))];
                     if (Session::haveRight('rssfeed_public', UPDATE)) {
                         if ($_SESSION['glpishow_count_on_tabs']) {
@@ -563,13 +441,8 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return '';
     }
 
-
-    /**
-     * @see CommonGLPI::defineTabs()
-     **/
     public function defineTabs($options = [])
     {
-
         $ong = [];
         $this->addDefaultFormTab($ong);
         $this->addStandardTab(__CLASS__, $ong, $options);
@@ -578,15 +451,8 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return $ong;
     }
 
-
-    /**
-     * @param $item         CommonGLPI object
-     * @param $tabnum       (default 1)
-     * @param $withtemplate (default 0)
-     **/
     public static function displayTabContentForItem(CommonGLPI $item, $tabnum = 1, $withtemplate = 0)
     {
-
         switch ($item->getType()) {
             case 'RSSFeed':
                 switch ($tabnum) {
@@ -653,23 +519,27 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
     private function checkUrlInput(string $url): bool
     {
         if (parse_url($url) === false) {
-            Session::addMessageAfterRedirect(__('Feed URL is invalid.'), false, ERROR);
+            Session::addMessageAfterRedirect(__s('Feed URL is invalid.'), false, ERROR);
             return false;
-        } elseif (!Toolbox::isUrlSafe($url)) {
-            Session::addMessageAfterRedirect(sprintf(__('URL "%s" is not allowed by your administrator.'), $url), false, ERROR);
+        }
+
+        if (!Toolbox::isUrlSafe($url)) {
+            Session::addMessageAfterRedirect(
+                htmlspecialchars(sprintf(__('URL "%s" is not allowed by your administrator.'), $url)),
+                false,
+                ERROR
+            );
             return false;
         }
 
         return true;
     }
 
-
     public function pre_updateInDB()
     {
-
-       // Set new user if initial user have been deleted
+        // Set new user if initial user have been deleted
         if (
-            ($this->fields['users_id'] == 0)
+            ((int) $this->fields['users_id'] === 0)
             && ($uid = Session::getLoginUserID())
         ) {
             $this->fields['users_id'] = $uid;
@@ -677,32 +547,28 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         }
     }
 
-
     public function post_getEmpty()
     {
-
         $this->fields["name"]         = __('New note');
         $this->fields["users_id"]     = Session::getLoginUserID();
         $this->fields["refresh_rate"] = DAY_TIMESTAMP;
         $this->fields["max_items"]    = 20;
     }
 
-
     /**
      * Print the rssfeed form
      *
-     * @param $ID        integer  Id of the item to print
-     * @param $options   array    of possible options:
+     * @param integer $ID Id of the item to print
+     * @param array $options Array of possible options:
      *     - target filename : where to go when done.
      **/
     public function showForm($ID, array $options = [])
     {
        // Test _rss cache directory. I permission trouble : unable to edit
         if (Toolbox::testWriteAccessToDirectory(GLPI_RSS_DIR) > 0) {
-            echo "<div class='center'>";
-            printf(__('Check permissions to the directory: %s'), GLPI_RSS_DIR);
-            echo "<p class='red b'>" . _n('Error', 'Errors', 1) . "</p>";
-            echo "</div>";
+            echo "<div class='alert alert-danger'><i class='alert-icon ti ti-alert-triangle'></i><div class='alert-title'>";
+            printf(__s('Check permissions to the directory: %s'), GLPI_RSS_DIR);
+            echo "</div></div>";
             return false;
         }
 
@@ -798,47 +664,44 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return true;
     }
 
-
     /**
      * Set error field
      *
-     * @param $error   (false by default
+     * @param boolean $error   (false by default)
      **/
     public function setError($error = false)
     {
-
         if (!isset($this->fields['id']) && !isset($this->fields['have_error'])) {
             return;
         }
 
-       // Set error if not set
+        // Set error if not set
         if ($error && !$this->fields['have_error']) {
-            $this->update(['id'         => $this->fields['id'],
+            $this->update([
+                'id'         => $this->fields['id'],
                 'have_error' => 1
             ]);
         }
-       // Unset error if set
+        // Unset error if set
         if (!$error && $this->fields['have_error']) {
-            $this->update(['id'         => $this->fields['id'],
+            $this->update([
+                'id'         => $this->fields['id'],
                 'have_error' => 0
             ]);
         }
     }
-
 
     /**
      * Show the feed content
      **/
     public function showFeedContent()
     {
-
         if (!$this->canViewItem()) {
             return false;
         }
         $rss_feed = [
             'items'  => []
         ];
-        echo "<div class='firstbloc'>";
         if ($feed = self::getRSSFeed($this->fields['url'], $this->fields['refresh_rate'])) {
             $this->setError(false);
             $rss_feed['title'] = $feed->get_title();
@@ -851,9 +714,9 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
                 ];
             }
         } else {
-            $rss_feed['error'] = !Toolbox::isUrlSafe($this->fields['url'])
+            $rss_feed['error'] = htmlspecialchars(!Toolbox::isUrlSafe($this->fields['url'])
                 ? sprintf(__('URL "%s" is not allowed by your administrator.'), $this->fields['url'])
-                : __('Error retrieving RSS feed');
+                : __('Error retrieving RSS feed'));
             $this->setError(true);
         }
 
@@ -861,7 +724,6 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
             'rss_feed'  => $rss_feed
         ]);
     }
-
 
     /**
      * Get a specific RSS feed.
@@ -914,7 +776,6 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         return $feed;
     }
 
-
     /**
      * Show list for central view
      *
@@ -943,7 +804,7 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
 
         if ($personal) {
            /// Personal notes only for central view
-            if (Session::getCurrentInterface() == 'helpdesk') {
+            if (Session::getCurrentInterface() === 'helpdesk') {
                 return false;
             }
 
@@ -958,14 +819,14 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
                 return false;
             }
 
-            $criteria = $criteria + self::getVisibilityCriteria();
+            $criteria += self::getVisibilityCriteria();
 
            // Only personal on central so do not keep it
-            if (Session::getCurrentInterface() == 'central') {
+            if (Session::getCurrentInterface() === 'central') {
                 $criteria['WHERE']["$table.users_id"] = ['<>', $users_id];
             }
 
-            if (Session::getCurrentInterface() == 'central') {
+            if (Session::getCurrentInterface() === 'central') {
                 $titre = "<a href=\"" . $CFG_GLPI["root_doc"] . "/front/rssfeed.php\">" .
                        _n('Public RSS feed', 'Public RSS feeds', Session::getPluralNumber()) . "</a>";
             } else {
@@ -1048,15 +909,9 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         }
     }
 
-    /**
-     * @since 0.85
-     *
-     * @see commonDBTM::getRights()
-     **/
     public function getRights($interface = 'central')
     {
-
-        if ($interface == 'helpdesk') {
+        if ($interface === 'helpdesk') {
             $values = [READ => __('Read')];
         } else {
             $values = parent::getRights();
@@ -1064,7 +919,6 @@ class RSSFeed extends CommonDBVisible implements ExtraVisibilityCriteria
         }
         return $values;
     }
-
 
     public static function getIcon()
     {